Issue-12 - Add support for uploading directly to an NG server
This change adds the ability to upload an EHSN file directly to an NG server.

It uses the AQTS python wrapper class to determine if the selected server is a 3.X server or an NG server.

If a 3.X server is detected, the same upload logic applies.
If an NG server is detected, the eHSN XML file and its PDF summary will be uploaded.